# Wasm marshaler API for user-defined types; migrate Uri and DateTime to new API