… src/libraries/System.Private.Runtime.InteropServices.JavaScript/src/ILLink/ILLink.Descriptors.LibraryBuild.xml
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/src/ILLink/ILLink.Descriptors.xml
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/src/System/Runtime/InteropServices/JavaScript/Runtime.cs
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/tests/ILLink.Descriptors.xml
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/tests/System.Private.Runtime.InteropServices.JavaScript.Tests.csproj
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/tests/System/Runtime/InteropServices/JavaScript/HelperMarshal.cs
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/tests/System/Runtime/InteropServices/JavaScript/JavaScriptTests.cs
… src/libraries/System.Private.Runtime.InteropServices.JavaScript/tests/System/Runtime/InteropServices/JavaScript/MarshalTests.cs
… src/mono/wasm/build/README.md
… src/mono/wasm/runtime/custom-marshaler.ts
… src/mono/wasm/runtime/dotnet.d.ts
… src/mono/wasm/runtime/driver.c
… src/mono/wasm/runtime/js-to-cs.ts
… src/mono/wasm/runtime/memory.ts
… src/mono/wasm/runtime/startup.ts
… src/mono/wasm/runtime/strings.ts
… src/mono/wasm/runtime/types.ts
… src/tasks/WasmAppBuilder/WasmAppBuilder.cs