This PR introduces a system for defining custom JS➔Managed and Managed➔JS marshaling primitives so that any* user-defined type can cross between environments seamlessly, and migrates our existing Uri and DateTime marshaling logic to use the new system.

First, some sample code:
```csharp
    public static class DateTimeMarshaler
    {
        public static string JavaScriptToInterchangeTransform => "return value.toISOString()";
        public static string InterchangeToJavaScriptTransform => "return new Date(value)";

        public static DateTime FromJavaScript (string s)
        {
            return DateTime.Parse(s).ToUniversalTime();
        }

        public static string ToJavaScript (in DateTime dt)
        {
            return dt.ToString("o");
        }
    }
```

General overview:
* The author of the type exposes static methods with specific names and signatures on the type so that the wasm bindings layer can find them.
  * Right now there is a manual linker configuration to keep our built-in marshalers from being linked out. The user will need to do this themselves for any marshalers they define.
  * The runtime receives a table of type➔marshaler mappings at startup, and this table is generated by WasmAppBuilder.
  * The ideal form of this in the future will be a source generator that automatically finds marshalers and generates the linker configuration + table, but right now this is not possible (source generators can't do the things necessary for this yet).
* The required methods are a JS➔Managed mapping and a Managed➔JS mapping. The types used by these methods have to (at present) be basic types that can cross through the bindings layer without assistance, i.e. double or string. In the future I hope to expand on this to allow marshaling spans or arrays.
* In addition to the direct JS⬌Managed mapping, you can optionally define additional methods that return JS expression filters. Those expression filters will be evaluated on the JS side of the boundary to process the value - for example since neither JS ```Date``` or C# ```DateTime``` can cross the bindings boundary directly, you can marshal it as a string and then use a filter to map it to/from the JS ```Date``` type.
  * TODO: It would be ideal to expose these as ```const string``` instead of function getters, but right now the only straightforward way to get at managed values is through function calls.
* Alongside all this we add a new ```'a'``` type specifier for method signatures (the strings you pass to ```call_static_method```, ```bind_static_method``` etc) that basically means 'figure it out'. When you use this specifier, the bindings layer will examine the target method and identify the best fit for that parameter. This is meant for the purpose of passing user-defined types so you don't need to think about whether they're classes or structs, but it also can be useful in other cases.
  * The downside is that any signature using this type specifier ends up being method-specific, so you will end up with method-specific generated code instead of the existing generated code that is shared by all methods with a given signature. There's room for some improvement here.

More detailed notes:
* Before this PR we had zero support for passing structs across the JS⬌Managed boundary in any circumstances (other than DateTime).
* The JS➔Managed conversion has to box the resulting value (if it's a struct), which introduces some unpleasant overhead. This ends up being less inefficient than you'd think (because currently we have to box *all* values being passed to managed methods from JS - not really sure why) but it's still an opportunity for performance improvement between eliminating some copies and ensuring the GC isn't involved.
* The introduction of support for marshaling custom classes introduces a bit of overhead for classes without custom marshaling implementations, because of the additional check added to the existing Managed➔JS flow. However, the new check is extremely cheap and the existing flow was very slow 😊 
* There are corner cases where this new system will not run (throwing a runtime exception) or will otherwise fail (for example if the signature of your methods is weird you can get garbage values). I think many of these can be tightened up if we write more test cases and add more error handling and checks to match.
* Many scenarios might call for an automated way to pass JSON blobs or raw byte data across the boundary. You can at least use filters (i.e. ```JSON.parse(value)```) to simplify the former, but the latter really demands integrated support for passing ```Span```s around and I have no idea how we'd do this in the existing setup.
* We could consider having some sort of default fallback marshaling implementation based on ```BinaryFormatter``` or something like that, but it seems out of scope. This system is at least relatively easy to extend to do that as a fallback.

Also in this PR:
* ```box_js_obj_with_converter``` API that allows you to request a specific managed type when boxing a JS value (this is how you create a type like Uri explicitly)
* Caching for the existing 'find class in assembly' APIs, so that they can be used efficiently at runtime when trying to box values of a specific type
* Caching for bind_method so that automated tests and benchmarks can safely bind methods at the start of each run without running out of runtime memory or incurring significant performance overhead (this was causing benchmarks to fail, because i had optimized the bindings so much that the run count in BDN was skyrocketing)
* Additional test coverage for some of the existing code, like being able to pass JS ```Date```s. I think that was actually broken, but Blazor works so whatever...
* Miscellaneous performance improvements, producing around a ~4-6% performance gain on my benchmarks for existing call paths
* Miscellaneous bug fixes and additional assertions in runtime/driver code (for things that previously would have produced crashes)
* Better debug information for generated code so it's easier to examine in browser debuggers and profilers