@kg 2021-01-29
To help with understanding some of what this machinery does, here's an example of some of the code generated by the bindings to facilitate calling C# from JS:
https://gist.github.com/kg/f2c350b007d8e783085124c5d03f508e