@radical 2021-02-02
This breaks debugger tests with `* Assertion: should not be reached at /Users/radical/dev/runtime/src/mono/mono/metadata/class-accessors.c:86`.
To reproduce: `$ make -C src/mono/wasm run-debugger-tests TEST_FILTER=DebuggerTests.ArrayTests.InvalidArrayId`

It fails at `this.async_method = Module.mono_bind_static_method ("[debugger-test] Math/NestedInMath:AsyncTest");`

The method in question: `public static async System.Threading.Tasks.Task<bool> AsyncTest(string s, int i)`
(https://github.com/dotnet/runtime/blob/master/src/mono/wasm/debugger/tests/debugger-test/debugger-test.cs#L139)

.. is in a nested class `Math.NestedInMath`.