@kg 2021-02-02
Thanks for digging in. I bet it's due to generics.