@kjpou1 2021-02-05
Not sure if it is generics or not but I had to add some code for handling generic signatures in the PR here: https://github.com/dotnet/runtime/pull/47519

This allows for generic signatures.