@kjpou1 2021-02-10
Just running through the DRAFT PR:

```
info: class_is_task
  info: class_is_task
  info: creating signature info for method result
  info: creating signature info for method params
  fail: [out of order message from the browser]: http://127.0.0.1:61482/dotnet.wasm 0:1703801 Uncaught RuntimeError: divide by zero
  info: GC_MINOR: (Nursery full) time 2.40ms, stw 2.44ms promoted 259K major size: 1280K in use: 687K los size: 0K in use: 0K
  info: GC_MINOR: (Nursery full) time 2.41ms, stw 2.43ms promoted 0K major size: 1280K in use: 687K los size: 0K in use: 0K
  info: GC_MINOR: (Nursery full) time 3.01ms, stw 3.03ms promoted 0K major size: 1280K in use: 687K los size: 0K in use: 0K
  info: GC_MINOR: (Nursery full) time 2.30ms, stw 2.32ms promoted 0K major size: 1280K in use: 687K los size: 0K in use: 0K
```