@kg 2021-05-22
> @kg are you still missing some changes?

I'm still working through the msbuild integration in a way that won't break tests