@kg 2021-05-27
> You can ignore the PGO failures. A newly added pipeline was enabled for PRs by mistake for a few hours.

Thanks, that reduces the number of failures to investigate.


I'm not sure why the new marshaler tests are failing on wasm CI, they work locally and I added the necessary msbuild configuration to feed in that data. I think something about the targets or build tasks isn't working - WasmAppBuilder should be feeding the config info into the runtime, and it's not.