@kg 2021-05-31
The MARSHAL_TYPE stuff for Uri/DateTime/DateTimeOffset remaining is intentional, if I remove those and a bunch of the other supporting structure, I break backwards compatibility. I have no way to know whether anyone is using that stuff. I'd be happy to remove it if we're OK with breaking existing code - I've been operating on the assumption we want existing code to work as-is. There are other changes I can make under those circumstances that will also improve things.