@kg 2021-06-05
OK, based on @lewing's concerns I added an explicit example and tests of marshaling complex types across the boundary, and made a couple changes to make it easier to do. Pre/post-filters require an explicit "return" now, but I don't feel like that's a problem since you get the ability to use local variables in exchange for it. I'll need to update the docs about this...