@kg 2021-06-12
I wrote some basic benchmarks to measure the impact of this PR on the speed of marshaling the built-in types that migrated over to the new system.

Before:
```
|                                          Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------ |----------:|----------:|----------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|                      InvokeMethodIntViaInvokeJS |  9.278 ms | 0.0733 ms | 0.0686 ms |  9.279 ms |  9.174 ms |  9.390 ms |     - |     - |     - |      18 B |
|                   InvokeMethodStringViaInvokeJS |  8.807 ms | 0.0426 ms | 0.0398 ms |  8.799 ms |  8.732 ms |  8.894 ms |     - |     - |     - |       2 B |
|           InvokeMethodStringViaInvokeJSWithArgs | 16.790 ms | 0.1052 ms | 0.0984 ms | 16.783 ms | 16.654 ms | 16.944 ms |     - |     - |     - | 160,076 B |
|              InvokeMethodIntViaInvokeJSWithArgs | 17.748 ms | 0.1283 ms | 0.1002 ms | 17.750 ms | 17.603 ms | 17.967 ms |     - |     - |     - | 160,110 B |
| InvokeMethodIntViaInvokeJSWithArgs_InternedName | 10.377 ms | 0.1115 ms | 0.1043 ms | 10.386 ms | 10.209 ms | 10.538 ms |     - |     - |     - | 160,216 B |
|              InvokeMethodUriViaInvokeJSWithArgs | 24.916 ms | 0.1868 ms | 0.1747 ms | 24.846 ms | 24.685 ms | 25.265 ms |     - |     - |     - | 160,082 B |
|             InvokeMethodDateViaInvokeJSWithArgs | 36.065 ms | 0.3771 ms | 0.3527 ms | 36.162 ms | 35.493 ms | 36.636 ms |     - |     - |     - | 320,106 B |
```

After:
```
|                                          Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------ |----------:|----------:|----------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|                      InvokeMethodIntViaInvokeJS |  9.173 ms | 0.0567 ms | 0.0502 ms |  9.164 ms |  9.099 ms |  9.295 ms |     - |     - |     - |      18 B |
|                   InvokeMethodStringViaInvokeJS |  8.895 ms | 0.0543 ms | 0.0481 ms |  8.893 ms |  8.829 ms |  8.991 ms |     - |     - |     - |       2 B |
|           InvokeMethodStringViaInvokeJSWithArgs | 10.065 ms | 0.0650 ms | 0.0608 ms | 10.060 ms |  9.954 ms | 10.167 ms |     - |     - |     - | 160,071 B |
|              InvokeMethodIntViaInvokeJSWithArgs | 10.763 ms | 0.0654 ms | 0.0612 ms | 10.774 ms | 10.661 ms | 10.857 ms |     - |     - |     - | 160,104 B |
| InvokeMethodIntViaInvokeJSWithArgs_InternedName | 10.862 ms | 0.0541 ms | 0.0480 ms | 10.877 ms | 10.770 ms | 10.934 ms |     - |     - |     - | 160,216 B |
|              InvokeMethodUriViaInvokeJSWithArgs | 17.148 ms | 0.1487 ms | 0.1318 ms | 17.148 ms | 16.970 ms | 17.374 ms |     - |     - |     - | 160,077 B |
|             InvokeMethodDateViaInvokeJSWithArgs | 28.815 ms | 0.1659 ms | 0.1386 ms | 28.801 ms | 28.619 ms | 29.032 ms |     - |     - |     - | 320,103 B |
```

The two "InvokeJS" entries are baseline tests just calling the test method(s) via running JS literals, and then the rest are using InvokeJSWithArgs, the API currently used by JSObject and pals to pass managed data to JS functions. It's a performance improvement for Uri and DateTime right now, but that might be entirely due to the interning fix - I'll have to rerun these measurements after the interning fix lands to see if there are other gains/losses then.